chore: new filters and updated allproperties File
