+++ app/(Home)/(Main Page)/Filters.tsx
@@ -1,5 +1,5 @@
 'use client';
-import { useState, type ReactElement } from 'react';
+import { useState, useEffect, type ReactElement } from 'react';
 
 import { ServerImage } from '@/components/shared/ServerImage';
 
@@ -8,9 +8,8 @@ interface Props {
 }
 
 export function Filters({ filterFunction }: Props): ReactElement {
-  const [categoryFilter, setCategoryFilter] = useState('');
-  const [countryFilter, setCountryFilter] = useState('');
-  const [annualYieldFilter, setAnnualYieldFilter] = useState('');
+  const [assetFilter, setAssetFilter] = useState('All');
+  const [subCategoryFilter, setSubCategoryFilter] = useState('All');
 
   const selectOptions: {
     imageURL: string;
@@ -19,35 +18,41 @@ export function Filters({ filterFunction }: Props): ReactElement {
   }[] = [
     {
       imageURL: '/assets/filter1.svg',
-      label: 'Property Type',
-      options: ['All', 'Residential', 'Commercial', 'Farm', 'Restaurant', 'Office'],
+      label: 'Asset',
+      options: ['All', 'Real Estate', 'Gems and Metals', 'Security', 'Art'],
     },
     {
       imageURL: '/assets/filter2.svg',
-      label: 'Location',
-      options: ['Worldwide', 'France', 'USA', 'Spain', 'Canada', 'Japan', 'United Kingdom', 'Netherlands', 'Brazil'],
-    },
-    {
-      imageURL: '/assets/filter3.svg',
-      label: 'Rental Yield',
-      options: ['Up to  5%', '5% to 10%', '10% to 15%', '15% to 30%'],
+      label: 'Subcategory',
+      options:
+        assetFilter === 'Real Estate'
+          ? ['All', 'Residential', 'Commercial', 'Farm', 'Restaurant', 'Office']
+          : assetFilter === 'Gems and Metals'
+            ? ['All', 'Gold', 'Emeralds', 'Diamonds']
+            : ['All'],
     },
   ];
 
-  const handleCategoryChange = (event: any) => {
-    setCategoryFilter(event.target.value);
+  useEffect(() => {
+    filterFunction(assetFilter, subCategoryFilter);
+  }, [assetFilter, subCategoryFilter]);
+
+  const handleAssetChange = (event: any) => {
+    setAssetFilter(event.target.value);
+    setSubCategoryFilter('All');
   };
 
-  const handleCountryChange = (event: any) => {
-    setCountryFilter(event.target.value);
+  const handleSubCategoryChange = (event: any) => {
+    setSubCategoryFilter(event.target.value);
   };
 
-  const handleAnnualYieldChange = (event: any) => {
-    setAnnualYieldFilter(event.target.value);
+  const resetFilters = () => {
+    setAssetFilter('All');
+    setSubCategoryFilter('All');
   };
 
   return (
-    <div className="flex flex-col align-start gap-2 self-stretch max-w-[95%] relative top-[-3rem] mb-[-1rem] p-2 z-[999] shadow-sm bg-white border lg:top-[-1.8rem] lg:flex-row lg:px-8 lg:py-6 justify-between lg:align-center lg:max-w-[75rem] m-auto rounded-[5px]">
+    <div className="flex flex-col align-start gap-8 self-stretch max-w-[95%] relative top-[-3rem] mb-[-1rem] p-2 z-[999] shadow-sm bg-white border lg:top-[-1.8rem] lg:flex-row lg:px-8 lg:py-6 justify-between lg:align-center lg:max-w-[75rem] m-auto rounded-[5px]">
       {selectOptions.map((option, index) => (
         <div key={index} className="w-full justify-between flex align-center md:gap-4 md:justify-center md:w-none">
           <div className="flex font-space_grotesk text-base font-bold justify-center items-center">
@@ -56,14 +61,10 @@ export function Filters({ filterFunction }: Props): ReactElement {
           </div>
 
           <select
-            className="flex w-[200px] cursor-pointer px-4 py-3 bg-white rounded-[5px] border border-black border-opacity-10"
-            onChange={
-              option.label === 'Property Type'
-                ? handleCategoryChange
-                : option.label === 'Location'
-                  ? handleCountryChange
-                  : handleAnnualYieldChange
-            }
+            className="flex w-full cursor-pointer px-4 py-3 bg-white rounded-[5px] border border-black border-opacity-10"
+            onChange={option.label === 'Asset' ? handleAssetChange : handleSubCategoryChange}
+            value={option.label === 'Asset' ? assetFilter : subCategoryFilter}
+            disabled={option.label === 'Subcategory' && selectOptions[1].options.length === 0}
           >
             {option.options.map((opt, idx) => (
               <option key={idx} value={opt}>
@@ -74,12 +75,9 @@ export function Filters({ filterFunction }: Props): ReactElement {
         </div>
       ))}
 
-      <button
-        className="flex bg-libertumGreen rounded-[5px] px-4 justify-center gap-4 py-4"
-        onClick={() => filterFunction(categoryFilter, countryFilter, annualYieldFilter)}
-      >
-        <p className="md:hidden text-white font-space_grotesk font-bold">Search Properties </p>
-        <ServerImage src="/assets/magnifying.svg" alt="N" width="30" height="30" className="md:w-[70px]" />
+      <button className="flex bg-libertumGreen rounded-[5px] px-4 justify-center gap-4 py-4" onClick={resetFilters}>
+        <p className="md:hidden text-white font-space_grotesk font-bold">Reset Filters</p>
+        <ServerImage src="/assets/reset.svg" alt="Reset" width="30" height="30" className="md:w-[70px]" />
       </button>
     </div>
   );
